Simplify `main` using functional programming stuff
Simplify `main` using functional programming stuff so that we don't have
so much noise going on.

This change hinges on the following changes and understandings.

The return type of the two functions being called are `Result<(), dyn
Error>`. Which matches the return type of our `main()`. Therefore, we
don't need to facilitate any conversion of errors, we can just promote
them up. However, if we wanted to add another layer with more specific
error naming, etc. we very well could.

The `build()` function returns a `Result<Config, dyn Error>`. So if we
know the wrapped value is a `Config` which is the key dependency for the
following call. We have the `map()` function on the `Result` type in the
Ok path unwraps the wrapped value, hands it to given function to
transform it. This is a standard function programming concept. However,
this isn't exactly what we want because if we use `map()` we will end up
with a `Result<Result, dyn Error>, dyn Error>` which no longer would
match the return type of `main()`. So we have to look for something
else. In this case it is the `bind()` function in most pure functional
programming languages, in other languages it is referred to as
`flat_map`, and others, like Rust it is referred to as `and_then()`.
This allows our transformation function to return a `Result<>` and it
conceptually handles the flattening of nesting of the `Result<>`.

It is also important that you know that if main returns a Result with a
std::error::Error then I am pretty sure it gets displayed to the user.
So you don't have to worry about doing the printing. You just have to
worry about making sure your Errors are defined well.

Other than that and probably least significant was that I just used
imports to make things shorter with out the namespacing in the calls.

<!-- ps-id: 82314b21-97a6-48a3-b85c-5fc77e19949c -->